libs/langchain_v1/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langchain"
description = "Building applications with LLMs through composability"
license = { text = "MIT" }
readme = "README.md"
authors = []

version = "1.1.3"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "langchain-core>=1.1.2,<2.0.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.2,<1.1.0",
    "pydantic>=2.7.4,<3.0.0",
]

[project.optional-dependencies]
community = ["langchain-community"]
anthropic = ["langchain-anthropic"]
openai = ["langchain-openai"]
azure-ai = ["langchain-azure-ai"]
#cohere = ["langchain-cohere"]
google-vertexai = ["langchain-google-vertexai"]
google-genai = ["langchain-google-genai"]
fireworks = ["langchain-fireworks"]
ollama = ["langchain-ollama"]
together = ["langchain-together"]
mistralai = ["langchain-mistralai"]
huggingface = ["langchain-huggingface"]
groq = ["langchain-groq"]
aws = ["langchain-aws"]
deepseek = ["langchain-deepseek"]
xai = ["langchain-xai"]
perplexity = ["langchain-perplexity"]

[project.urls]
Homepage = "https://docs.langchain.com/"
Documentation = "https://reference.langchain.com/python/langchain/langchain/"
Source = "https://github.com/langchain-ai/langchain/tree/master/libs/langchain"
Changelog = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain%3D%3D1%22"
Twitter = "https://x.com/LangChainAI"
Slack = "https://www.langchain.com/join-community"
Reddit = "https://www.reddit.com/r/LangChain/"

[dependency-groups]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=4.0.0,<8.0.0",
    "pytest-watcher>=0.2.6,<1.0.0",
    "pytest-asyncio>=0.23.2,<2.0.0",
    "pytest-socket>=0.6.0,<1.0.0",
    "pytest-xdist<4.0.0,>=3.6.1",
    "pytest-mock",
    "syrupy>=4.0.2,<5.0.0",
    "toml>=0.10.2,<1.0.0",
    "langchain-tests",
    "langchain-openai",
]
lint = [
    "ruff>=0.12.2,<0.13.0",
]
typing = [
    "mypy>=1.18.1,<1.19.0",
    "types-toml>=0.10.8.20240310,<1.0.0.0",
]

test_integration = [
    "vcrpy>=7.0.0,<8.0.0",
    "wrapt>=1.15.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "cassio>=0.1.0,<1.0.0",
    "langchainhub>=0.1.16,<1.0.0",
    "langchain-core",
    "langchain-text-splitters",
]

[tool.uv]
prerelease = "allow"

[tool.uv.sources]
langchain-core = { path = "../core", editable = true }
langchain-tests = { path = "../standard-tests", editable = true }
langchain-text-splitters = { path = "../text-splitters", editable = true }
langchain-openai = { path = "../partners/openai", editable = true }

[tool.ruff]
exclude = ["tests/integration_tests/examples/non-utf8-encoding.py"]
line-length = 100

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = "deprecated"
exclude = ["tests/unit_tests/agents/*", "tests/integration_tests/agents/*"]

# TODO: activate for 'strict' checking
disallow_any_generics = false
warn_return_any = false

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "ALL"
]
ignore = [
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "SLF001",  # Private member access
    "PLC0415", # Imports should be at the top. Not always desirable
    "PLR0913", # Too many arguments in function definition
    "PLC0414", # Inconsistent with how type checkers expect to be notified of intentional re-exports
    "RUF002",  # Em-dash in docstring
]
unfixable = ["B028"] # People should intentionally tune the stacklevel

pyupgrade.keep-runtime-typing = true
flake8-annotations.allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-var-parameters = true  # ignore missing documentation for *args and **kwargs parameters

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D1", # Documentation rules
    "PLC0415", # Imports should be at the top. Not always desirable for tests
]
"langchain/agents/*" = [
    "ANN401", # we use Any right now, need to narrow
    "A002", # input is shadowing builtin
    "A001", # input is shadowing builtin
    "B904", # use from for exceptions
    "PLR2004", # magic values are fine for this case
    "C901", # too complex
    "TRY004", # type error exception
    "PLR0912", # too many branches
    "PLR0911", # too many return statements
]
"tests/unit_tests/agents/*" = ["ALL"]
"tests/integration_tests/agents/*" = ["ALL"]

[tool.ruff.lint.extend-per-file-ignores]
"scripts/check_imports.py" = ["ALL"]

"langchain/chat_models/base.py" = [
    "ANN",
    "C901",
    "FIX002",
    "N802",
    "PLR0911",
    "PLR0912",
    "PLR0915",
]

"langchain/embeddings/base.py" = [
    "PLR0911",
    "PLR0913",
]

"tests/**/*.py" = [
    "S101",   # Tests need assertions
    "S311",   # Standard pseudo-random generators are not suitable for cryptographic purposes
    "SLF001", # Private member access in tests
    "PLR2004", # Magic values are perfectly fine in unit tests (e.g. 0, 1, 2, etc.)
    "C901", # Too complex
    "ANN401", # Annotated type is not necessary
    "N802", # Function name should be lowercase
    "PLW1641", # Object does not implement __hash__ method
    "ARG002", # Unused argument
    "BLE001", # Do not catch blind exception
    "N801", # class name should use CapWords convention
]

"langchain/tools/tool_node.py" = [
    "ANN401", # we use Any right now, need to narrow
    "A002", # input is shadowing builtin
    "A001", # input is shadowing builtin
    "B904", # use from for exceptions
    "PLR2004", # magic values are fine for this case
    "C901", # too complex
    "TRY004", # type error exception
]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5 --snapshot-warn-unused -vv"
markers = [
    "requires: mark tests as requiring a specific library",
    "scheduled: mark tests to run in scheduled testing",
    "compile: mark placeholder test used to compile integration tests without running them",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::langchain_core._api.beta_decorator.LangChainBetaWarning",
    "ignore::langchain_core._api.deprecation.LangChainDeprecationWarning:tests",
    "ignore::langchain_core._api.deprecation.LangChainPendingDeprecationWarning:tests",
]
